+++ nengo_loihi/emulator/interface.py
@@ -82,10 +82,7 @@ class EmulatorInterface(object):
         self.synapses = None
         self.axons = None
 
-    def chip2host(self, probes_receivers=None):
-        if probes_receivers is None:
-            probes_receivers = {}
-
+    def chip2host(self, probes_receivers):
         increment = 0
         for probe, receiver in probes_receivers.items():
             inc = self.probes.send(probe, self._chip2host_sent_steps, receiver)
@@ -210,19 +207,6 @@ class IterableState(object):
                 # multiple items per block (attribute is iterable)
                 for item in getattr(block, block_key):
                     yield block, item
-
-    def __contains__(self, item):
-        return item in self.slices
-
-    def __getitem__(self, key):
-        return self.slices[key]
-
-    def __iter__(self):
-        for obj in self.slices:
-            yield obj
-
-    def __len__(self):
-        return len(self.slices)
 
     def error(self, msg):
         if self.strict: