+++ nengo_loihi/hardware/interface.py
@@ -123,12 +123,6 @@ class HardwareInterface(object):
         # --- build
         self.n2board = build_board(self.board)
 
-    def print_cores(self):
-        for j, n2chip in enumerate(self.n2board.n2Chips):
-            print("Chip %d, id=%d" % (j, n2chip.id))
-            for k, n2core in enumerate(n2chip.n2Cores):
-                print("  Core %d, id=%d" % (k, n2core.id))
-
     def run_steps(self, steps, blocking=True):
         if self.use_snips and self.nengo_io_h2c is None:
             self.create_io_snip()
@@ -198,10 +192,7 @@ class HardwareInterface(object):
 
         self._chip2host_sent_steps += 1
 
-    def chip2host(self, probes_receivers=None):
-        if probes_receivers is None:
-            probes_receivers = {}
-
+    def chip2host(self, probes_receivers):
         return (self._chip2host_snips(probes_receivers) if self.use_snips else
                 self._chip2host_monitor(probes_receivers))